docs: refactor equipment items into their own module
